chore(lint): adopt ruff 0.16 with a deliberate config pass (#183) (#186)
Why
- Dependabot's ruff 0.16 bump was declined as a drive-by (619 flagged
  sites); this is the deliberate pass the decline promised, so the
  toolchain stops aging.

What changed
- pyproject: ruff>=0.16,<0.17 with justified ignores — B008 (FastAPI
  Depends-in-default idiom), DTZ011/DTZ007 (calendar days and calendar
  string parsing are deliberate), C408 (keyword-dict row building).
- 99 autofixes (import sorting, quoted annotations, stale noqas, regex
  flag aliases) + hand fixes: unnecessary int casts, zip->pairwise,
  collapsed nested ifs, parenthesized implicit string concats, narrowed
  a blind pytest.raises, reasoned noqas for tiny sync writes in upload
  handlers and final-safety-net excepts.
- Real bug found by the pass: three ai_runtime endpoints declared
  async def with NO awaits while making blocking HTTP calls
  (multi-second HF metadata timeout) — they stalled the event loop for
  every concurrent request. Now def, so FastAPI runs them in the
  threadpool.

Verification
- apps/api 691 passed both with and without a master key;
  check-openapi clean; ruff 0.16.1 zero findings.

Sensitive data check
- Tooling and mechanical fixes only.

+++ apps/api/src/family_cfo_api/api/accounts.py
@@ -129,7 +129,7 @@ def parse_account_scan(text: str) -> AccountScanResult:
 
     def money_minor(key: str) -> int | None:
         value = _scan_number(data.get(key))
-        return int(round(value * 100)) if value is not None and value > 0 else None
+        return round(value * 100) if value is not None and value > 0 else None
 
     cash = money_minor("cash_balance")
     investment = money_minor("investment_value")
@@ -263,8 +263,8 @@ def parse_loan_statement_text(text: str) -> LoanScanResult | None:
         ],
     )
     return LoanScanResult(
-        monthly_payment_minor=int(round(payment * 100)) if payment and payment > 0 else None,
-        balance_minor=int(round(balance * 100)) if balance and balance > 0 else None,
+        monthly_payment_minor=round(payment * 100) if payment and payment > 0 else None,
+        balance_minor=round(balance * 100) if balance and balance > 0 else None,
         next_payment_due_date=_parse_label_date(due) if due else None,
         apr_percent=rate if rate is not None and 0 <= rate < 100 else None,
         note="Read from the statement text — confirm every value before saving.",
@@ -294,7 +294,7 @@ def parse_loan_scan(text: str) -> LoanScanResult:
 
     def money_minor(key: str) -> int | None:
         value = _scan_number(data.get(key))
-        return int(round(value * 100)) if value is not None and value > 0 else None
+        return round(value * 100) if value is not None and value > 0 else None
 
     def positive_int(key: str) -> int | None:
         value = _scan_number(data.get(key))
